refactor!: prove `iIndepFun.prod` in terms of `iIndepFun.pi'`
This replaces a `Nonempty` assumption with a `Fintype` assumption. Is
this OK? i.e. if we need independence of an infinite family of finite
sets of random variables, we might not be able to use `iIndepFun.pi'`
(as is currently).

I'm not familiar enough with math to know if the `Fintype` assumption is
necessary to prove `iIndepFun.pi` in "real life"; in our Lean proof, it
is used to apply `iIndepFun_iff_pi_map_eq_map`.